Optimized
Instead of using a queue to store the audio data, you can directly pass the data to the KaldiRecognizer in the callback function. This can reduce the overhead of managing the queue.

You can increase the blocksize of the RawInputStream to reduce the number of callback invocations, which can improve performance.

In this code, the audio data is passed directly to the KaldiRecognizer in the callback_wrapper function. The blocksize of the RawInputStream is increased to match the size of the audio data that is passed to the recognizer in each callback. Finally, an infinite loop is added to keep the program running